feat: implementações de arquivos locais e deletando aruqivos de build/IDE Eclipse

+++ mercadinho-providence/src/main/java/br/com/mercadinhoprovidence/model/LoteEstoque.java
@@ -19,7 +19,7 @@ public class LoteEstoque {
     private Integer idProduto;
     private BigDecimal precoCusto;
     private LocalDate validade;
-    private Double quantidadeAtual;
+    private BigDecimal quantidadeAtual;
     private LocalDateTime dataEntrada;
     private Boolean ativo;
